responsive styles for categories list done
