+++ src/Styles/TablesStyles.css
@@ -17,3 +17,7 @@
     font-size: 1rem;
   }
 }
+
+.MuiTablePagination-toolbar p {
+  display: block;
+}
